+++ src/Contracts/Requests/UpdateMasterProductVariant.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the UpdateMasterProductVariant interface.
+ *
+ * @copyright   Copyright (c) 2022 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2022-12-20
+ *
+ */
+
+namespace Vanilo\Admin\Contracts\Requests;
+
+use Konekt\Concord\Contracts\BaseRequest;
+
+interface UpdateMasterProductVariant extends BaseRequest
+{
+}