+++ src/resources/routes/admin.php
@@ -36,3 +36,9 @@ Route::get('/property/{property}/value/{property_value}/edit', 'PropertyValueCon
 Route::put('/property/{property}/value/{property_value}', 'PropertyValueController@update')->name('property_value.update');
 Route::delete('/property/{property}/value/{property_value}', 'PropertyValueController@destroy')->name('property_value.destroy');
 Route::put('/property/sync/{for}/{forId}', 'PropertyValueController@sync')->name('property_value.sync');
+
+Route::get('/master-product/{masterProduct}/variant/create', 'MasterProductVariantController@create')->name('master-product-variant.create');
+Route::post('/master-product/{masterProduct}/variant', 'MasterProductVariantController@store')->name('master-product-variant.store');
+Route::get('/master-product/{masterProduct}/variant/{masterProductVariant}/edit', 'MasterProductVariantController@edit')->name('master-product-variant.edit');
+Route::put('/master-product/{masterProduct}/variant/{masterProductVariant}', 'MasterProductVariantController@update')->name('master-product-variant.update');
+//Route::delete('/taxonomy/{taxonomy}/taxon/{taxon}', 'TaxonController@destroy')->name('taxon.destroy');