+++ src/resources/routes/breadcrumbs.php
@@ -37,6 +37,16 @@ Breadcrumbs::for('vanilo.admin.master-product.edit', function ($breadcrumbs, $pr
     $breadcrumbs->push(__('Edit'), route('vanilo.admin.master-product.edit', $product));
 });
 
+Breadcrumbs::for('vanilo.admin.master-product-variant.create', function ($breadcrumbs, $masterProduct) {
+    $breadcrumbs->parent('vanilo.admin.master-product.show', $masterProduct);
+    $breadcrumbs->push(__('Create Variant'));
+});
+
+Breadcrumbs::for('vanilo.admin.master-product-variant.edit', function ($breadcrumbs, $masterProduct, $masterProductVariant) {
+    $breadcrumbs->parent('vanilo.admin.master-product.show', $masterProduct);
+    $breadcrumbs->push($masterProductVariant->name);
+});
+
 Breadcrumbs::for('vanilo.admin.taxonomy.index', function ($breadcrumbs) {
     $breadcrumbs->parent('home');
     $breadcrumbs->push(__('Categorization'), route('vanilo.admin.taxonomy.index'));